refactor: Split switch-account submenu into trigger and content
Remove the old SwitchAccountItem and introduce a dedicated SwitchAccountSubmenuContent component that contains the session list, current session indicator, and add-account link. Update SwitchAccountSubmenu to act as the submenu trigger (accepting an optional className) and render the new content component inside a DropdownMenuSub. Update the multi-session plugin to register the SwitchAccountSubmenu as the user menu item. This refactor defers the device sessions query to the submenu content so fetching is only performed when the submenu is opened and simplifies the menu component structure.
