+++ examples/start-shadcn-example/src/components/user/switch-account-submenu-content.tsx
@@ -1,0 +1,69 @@
+"use client"
+
+import {
+  type MultiSessionAuthClient,
+  useAuth,
+  useAuthPlugin,
+  useListDeviceSessions,
+  useSession
+} from "@better-auth-ui/react"
+import { Check, CirclePlus } from "lucide-react"
+
+import {
+  DropdownMenuItem,
+  DropdownMenuSeparator,
+  DropdownMenuSubContent
+} from "@/components/ui/dropdown-menu"
+import { multiSessionPlugin } from "@/lib/multi-session/multi-session-plugin"
+import { SwitchAccountSubmenuItem } from "./switch-account-submenu-item"
+import { UserView } from "./user-view"
+
+/**
+ * Render the submenu content for switching between multiple authenticated sessions.
+ *
+ * Shows the current session with a checkmark, lists other device sessions that can be activated,
+ * and provides an option to add a new account. This component should be rendered inside a
+ * DropdownMenuSub to defer the useListDeviceSessions query until the submenu is opened.
+ *
+ * @returns The switch account submenu content as a JSX element
+ */
+export function SwitchAccountSubmenuContent() {
+  const { authClient, basePaths, viewPaths, Link } = useAuth()
+  const { localization: multiSessionLocalization } =
+    useAuthPlugin(multiSessionPlugin)
+  const { data: session } = useSession(authClient)
+  const { data: deviceSessions, isPending } = useListDeviceSessions(
+    authClient as MultiSessionAuthClient
+  )
+
+  return (
+    <DropdownMenuSubContent className="min-w-48 md:min-w-56 max-w-[48svw]">
+      <DropdownMenuItem>
+        <UserView isPending={isPending} />
+
+        {!isPending && <Check className="ml-auto" />}
+      </DropdownMenuItem>
+
+      {deviceSessions
+        ?.filter(
+          (deviceSession) => deviceSession.session.id !== session?.session.id
+        )
+        .map((deviceSession) => (
+          <SwitchAccountSubmenuItem
+            key={deviceSession.session.id}
+            deviceSession={deviceSession}
+          />
+        ))}
+
+      <DropdownMenuSeparator />
+
+      <DropdownMenuItem asChild>
+        <Link href={`${basePaths.auth}/${viewPaths.auth.signIn}`}>
+          <CirclePlus className="text-muted-foreground" />
+
+          {multiSessionLocalization.addAccount}
+        </Link>
+      </DropdownMenuItem>
+    </DropdownMenuSubContent>
+  )
+}
