+++ examples/start-shadcn-example/src/lib/multi-session/multi-session-plugin.ts
@@ -4,13 +4,13 @@ import {
 } from "@better-auth-ui/core/plugins"
 
 import { ManageAccounts } from "@/components/settings/account/manage-accounts"
-import { SwitchAccountItem } from "@/components/user/switch-account-item"
+import { SwitchAccountSubmenu } from "@/components/user/switch-account-submenu"
 import type { AuthPlugin } from "../auth-plugin"
 
 export function multiSessionPlugin(options: MultiSessionPluginOptions = {}) {
   return {
     ...coreMultiSessionPlugin(options),
     accountCards: [ManageAccounts],
-    userMenuItems: [SwitchAccountItem]
+    userMenuItems: [SwitchAccountSubmenu]
   } satisfies AuthPlugin
 }
